fix: asymmetric weights in optimization (#17)
Fixes the asymmetric weighting calculation during optimization. 

Previously, the weighting seemed to use the inverse of the limits for
the score ratio as the weight. Now it uses "estimated variance".

The asymmetric weighting means that the positive errors are allowed to
be larger than negative errors. This is because score ratios have hard
minimum at zero and negative error size must be limited; the loss
function is asymmetric.

This hasn't yet been proved to be a good idea in practice. If this does
not work, one idea is to use same weights for positive and negative
errors. Need to investigate this a bit more.